Fix ClangCL errors and warnings in Nabla and tools

+++ include/nbl/asset/utils/CQuantNormalCache.h
@@ -42,8 +42,8 @@ struct QuantNormalHash
 		static constexpr size_t primeNumber1 = 18446744073709551557ull;
 		static constexpr size_t primeNumber2 = 4611686018427388273ull;
 				
-		return  ((static_cast<size_t>(static_cast<double>(vec.u)*(std::numeric_limits<size_t>::max)()) * primeNumber1) ^
-			(static_cast<size_t>(static_cast<double>(vec.v)*(std::numeric_limits<size_t>::max)()) * primeNumber2));
+		return  ((static_cast<size_t>(static_cast<double>(vec.u)*static_cast<double>((std::numeric_limits<size_t>::max)())) * primeNumber1) ^
+			(static_cast<size_t>(static_cast<double>(vec.v)*static_cast<double>((std::numeric_limits<size_t>::max)())) * primeNumber2));
 	}
 };
